fast_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic>=2.0
pyodbc
python-dotenv
